+++ CMakeLists.txt
@@ -3,5 +3,5 @@ project(Final)
 
 set(CMAKE_C_STANDARD 99)
 
-set(SOURCE_FILES projet/main.c projet/csv.c projet/csv.h projet/account.h projet/customer.h)
+set(SOURCE_FILES projet/main.c projet/csv.c projet/csv.h projet/account.h projet/customer.h projet/account.c)
 add_executable(Final ${SOURCE_FILES})
